+++ cmd/in.go
@@ -51,15 +51,13 @@ var inCmd = &cobra.Command{
 			return err
 		}
 
-		format, _ := cmd.Flags().GetString("format")
-		asJSON, _ := cmd.Flags().GetBool("json")
 		return manageEntry(
 			c,
 			tei,
 			createTimeEntry(c, true),
 			viper.GetBool(INTERACTIVE),
 			viper.GetBool(ALLOW_PROJECT_NAME),
-			printTimeEntryImpl(c, format, asJSON),
+			printTimeEntryImpl(c, cmd),
 			!viper.GetBool(ALLOW_INCOMPLETE),
 			true,
 		)